Bug Fixes
- Improving rendering of SVG images
- Allow markers on lines, polygons, and polylines
- Implement different line joins and line caps
- Improve CSS handling
