+++ Source/Basic Shapes/SvgImage.cs
@@ -109,88 +109,116 @@ namespace Svg
 
             if (Width.Value > 0.0f && Height.Value > 0.0f && this.Href != null)
             {
-                using (Image b = GetImage(this.Href))
+                var img = GetImage(this.Href);
+                if (img != null)
                 {
-                    if (b != null)
+                    RectangleF srcRect;
+                    var bmp = img as Image;
+                    var svg = img as SvgFragment;
+                    if (bmp != null)
                     {
-                        var srcRect = new RectangleF(0, 0, b.Width, b.Height);
-                        var destClip = new RectangleF(this.Location.ToDeviceValue(renderer, this),
-                                        new SizeF(Width.ToDeviceValue(renderer, UnitRenderingType.Horizontal, this), 
-                                                  Height.ToDeviceValue(renderer, UnitRenderingType.Vertical, this)));
-                        RectangleF destRect = destClip;
-                        
-                        this.PushTransforms(renderer);
-                        renderer.SetClip(new Region(destClip), CombineMode.Intersect);
-                        this.SetClip(renderer);
-
-                        if (AspectRatio != null && AspectRatio.Align != SvgPreserveAspectRatio.none)
-                        {
-                            var fScaleX = destClip.Width / srcRect.Width;
-                            var fScaleY = destClip.Height / srcRect.Height;
-                            var xOffset = 0.0f;
-                            var yOffset = 0.0f;
-
-                            if (AspectRatio.Slice)
-                            {
-                                fScaleX = Math.Max(fScaleX, fScaleY);
-                                fScaleY = Math.Max(fScaleX, fScaleY);
-                            }
-                            else
-                            {
-                                fScaleX = Math.Min(fScaleX, fScaleY);
-                                fScaleY = Math.Min(fScaleX, fScaleY);
-                            }
-
-                            switch (AspectRatio.Align)
-                            {
-                                case SvgPreserveAspectRatio.xMinYMin:
-                                    break;
-                                case SvgPreserveAspectRatio.xMidYMin:
-                                    xOffset = (destClip.Width - srcRect.Width * fScaleX) / 2;
-                                    break;
-                                case SvgPreserveAspectRatio.xMaxYMin:
-                                    xOffset = (destClip.Width - srcRect.Width * fScaleX);
-                                    break;
-                                case SvgPreserveAspectRatio.xMinYMid:
-                                    yOffset = (destClip.Height - srcRect.Height * fScaleY) / 2;
-                                    break;
-                                case SvgPreserveAspectRatio.xMidYMid:
-                                    xOffset = (destClip.Width - srcRect.Width * fScaleX) / 2;
-                                    yOffset = (destClip.Height - srcRect.Height * fScaleY) / 2;
-                                    break;
-                                case SvgPreserveAspectRatio.xMaxYMid:
-                                    xOffset = (destClip.Width - srcRect.Width * fScaleX);
-                                    yOffset = (destClip.Height - srcRect.Height * fScaleY) / 2;
-                                    break;
-                                case SvgPreserveAspectRatio.xMinYMax:
-                                    yOffset = (destClip.Height - srcRect.Height * fScaleY);
-                                    break;
-                                case SvgPreserveAspectRatio.xMidYMax:
-                                    xOffset = (destClip.Width - srcRect.Width * fScaleX) / 2;
-                                    yOffset = (destClip.Height - srcRect.Height * fScaleY);
-                                    break;
-                                case SvgPreserveAspectRatio.xMaxYMax:
-                                    xOffset = (destClip.Width - srcRect.Width * fScaleX);
-                                    yOffset = (destClip.Height - srcRect.Height * fScaleY);
-                                    break;
-                            }
-
-                            destRect = new RectangleF(destClip.X + xOffset, destClip.Y + yOffset, 
-                                                      srcRect.Width * fScaleX, srcRect.Height * fScaleY);
-                        }
-                        
-                        renderer.DrawImage(b, destRect, srcRect, GraphicsUnit.Pixel);
-
-                        this.ResetClip(renderer);
-                        this.PopTransforms(renderer);
+                        srcRect = new RectangleF(0, 0, bmp.Width, bmp.Height);
                     }
+                    else if (svg != null)
+                    {
+                        srcRect = new RectangleF(new PointF(0, 0), svg.GetDimensions());
+                    }
+                    else
+                    {
+                        return;
+                    }
+
+                    var destClip = new RectangleF(this.Location.ToDeviceValue(renderer, this),
+                                                  new SizeF(Width.ToDeviceValue(renderer, UnitRenderingType.Horizontal, this), 
+                                                            Height.ToDeviceValue(renderer, UnitRenderingType.Vertical, this)));
+                    RectangleF destRect = destClip;
+                        
+                    this.PushTransforms(renderer);
+                    renderer.SetClip(new Region(destClip), CombineMode.Intersect);
+                    this.SetClip(renderer);
+
+                    if (AspectRatio != null && AspectRatio.Align != SvgPreserveAspectRatio.none)
+                    {
+                        var fScaleX = destClip.Width / srcRect.Width;
+                        var fScaleY = destClip.Height / srcRect.Height;
+                        var xOffset = 0.0f;
+                        var yOffset = 0.0f;
+
+                        if (AspectRatio.Slice)
+                        {
+                            fScaleX = Math.Max(fScaleX, fScaleY);
+                            fScaleY = Math.Max(fScaleX, fScaleY);
+                        }
+                        else
+                        {
+                            fScaleX = Math.Min(fScaleX, fScaleY);
+                            fScaleY = Math.Min(fScaleX, fScaleY);
+                        }
+
+                        switch (AspectRatio.Align)
+                        {
+                            case SvgPreserveAspectRatio.xMinYMin:
+                                break;
+                            case SvgPreserveAspectRatio.xMidYMin:
+                                xOffset = (destClip.Width - srcRect.Width * fScaleX) / 2;
+                                break;
+                            case SvgPreserveAspectRatio.xMaxYMin:
+                                xOffset = (destClip.Width - srcRect.Width * fScaleX);
+                                break;
+                            case SvgPreserveAspectRatio.xMinYMid:
+                                yOffset = (destClip.Height - srcRect.Height * fScaleY) / 2;
+                                break;
+                            case SvgPreserveAspectRatio.xMidYMid:
+                                xOffset = (destClip.Width - srcRect.Width * fScaleX) / 2;
+                                yOffset = (destClip.Height - srcRect.Height * fScaleY) / 2;
+                                break;
+                            case SvgPreserveAspectRatio.xMaxYMid:
+                                xOffset = (destClip.Width - srcRect.Width * fScaleX);
+                                yOffset = (destClip.Height - srcRect.Height * fScaleY) / 2;
+                                break;
+                            case SvgPreserveAspectRatio.xMinYMax:
+                                yOffset = (destClip.Height - srcRect.Height * fScaleY);
+                                break;
+                            case SvgPreserveAspectRatio.xMidYMax:
+                                xOffset = (destClip.Width - srcRect.Width * fScaleX) / 2;
+                                yOffset = (destClip.Height - srcRect.Height * fScaleY);
+                                break;
+                            case SvgPreserveAspectRatio.xMaxYMax:
+                                xOffset = (destClip.Width - srcRect.Width * fScaleX);
+                                yOffset = (destClip.Height - srcRect.Height * fScaleY);
+                                break;
+                        }
+
+                        destRect = new RectangleF(destClip.X + xOffset, destClip.Y + yOffset, 
+                                                    srcRect.Width * fScaleX, srcRect.Height * fScaleY);
+                    }
+
+                    if (bmp != null)
+                    {
+                        renderer.DrawImage(bmp, destRect, srcRect, GraphicsUnit.Pixel);
+                        bmp.Dispose();
+                    }
+                    else if (svg != null)
+                    {
+                        var currOffset = new PointF(renderer.Transform.OffsetX, renderer.Transform.OffsetY);
+                        renderer.TranslateTransform(-currOffset.X, -currOffset.Y);
+                        renderer.ScaleTransform(destRect.Width / srcRect.Width, destRect.Height / srcRect.Height);
+                        renderer.TranslateTransform(currOffset.X + destRect.X, currOffset.Y + destRect.Y);
+                        renderer.SetBoundable(new GenericBoundable(srcRect));
+                        svg.RenderElement(renderer);
+                        renderer.PopBoundable();
+                    }
+
+                    
+                    this.ResetClip(renderer);
+                    this.PopTransforms(renderer);
                 }
                 // TODO: cache images... will need a shared context for this
                 // TODO: support preserveAspectRatio, etc
             }
         }
 
-        protected Image GetImage(Uri uri)
+        protected object GetImage(Uri uri)
         {
             try
             {
@@ -228,7 +256,7 @@ namespace Svg
                         {
                             var doc = SvgDocument.Open<SvgDocument>(stream);
                             doc.BaseUri = uri;
-                            return doc.Draw();
+                            return doc;
                         }
                         else
                         {
