+++ Source/Basic Shapes/SvgPolygon.cs
@@ -15,16 +15,47 @@ namespace Svg
     public class SvgPolygon : SvgVisualElement
     {
         private GraphicsPath _path;
-        private SvgPointCollection _points;
-
+        
         /// <summary>
         /// The points that make up the SvgPolygon
         /// </summary>
         [SvgAttribute("points")]
         public SvgPointCollection Points
         {
-            get { return this._points; }
-            set { this._points = value; this.IsPathDirty = true; }
+            get { return this.Attributes["points"] as SvgPointCollection; }
+            set { this.Attributes["points"] = value; this.IsPathDirty = true; }
+        }
+
+        /// <summary>
+        /// Gets or sets the marker (end cap) of the path.
+        /// </summary>
+        [SvgAttribute("marker-end")]
+        public Uri MarkerEnd
+        {
+            get { return this.Attributes.GetAttribute<Uri>("marker-end"); }
+            set { this.Attributes["marker-end"] = value; }
+        }
+
+
+        /// <summary>
+        /// Gets or sets the marker (start cap) of the path.
+        /// </summary>
+        [SvgAttribute("marker-mid")]
+        public Uri MarkerMid
+        {
+            get { return this.Attributes.GetAttribute<Uri>("marker-mid"); }
+            set { this.Attributes["marker-mid"] = value; }
+        }
+
+
+        /// <summary>
+        /// Gets or sets the marker (start cap) of the path.
+        /// </summary>
+        [SvgAttribute("marker-start")]
+        public Uri MarkerStart
+        {
+            get { return this.Attributes.GetAttribute<Uri>("marker-start"); }
+            set { this.Attributes["marker-start"] = value; }
         }
 
         protected override bool RequiresSmoothRendering
@@ -41,14 +72,15 @@ namespace Svg
 
                 try
                 {
-                    for (int i = 2; (i + 1) < this._points.Count; i += 2)
+                    var points = this.Points;
+                    for (int i = 2; (i + 1) < points.Count; i += 2)
                     {
-                        var endPoint = SvgUnit.GetDevicePoint(this._points[i], this._points[i+1], renderer, this);
+                        var endPoint = SvgUnit.GetDevicePoint(points[i], points[i + 1], renderer, this);
 
                         //first line
                         if (_path.PointCount == 0)
                         {
-                            _path.AddLine(SvgUnit.GetDevicePoint(this._points[i - 2], this._points[i - 1], renderer, this), endPoint);
+                            _path.AddLine(SvgUnit.GetDevicePoint(points[i - 2], points[i - 1], renderer, this), endPoint);
                         }
                         else
                         {
@@ -65,6 +97,37 @@ namespace Svg
                 this.IsPathDirty = false;
             }
             return this._path;
+        }
+
+        /// <summary>
+        /// Renders the stroke of the <see cref="SvgVisualElement"/> to the specified <see cref="ISvgRenderer"/>
+        /// </summary>
+        /// <param name="renderer">The <see cref="ISvgRenderer"/> object to render to.</param>
+        protected internal override bool RenderStroke(ISvgRenderer renderer)
+        {
+            var result = base.RenderStroke(renderer);
+            var path = this.Path(renderer);
+
+            if (this.MarkerStart != null)
+            {
+                SvgMarker marker = this.OwnerDocument.GetElementById<SvgMarker>(this.MarkerStart.ToString());
+                marker.RenderMarker(renderer, this, path.PathPoints[0], path.PathPoints[0], path.PathPoints[1]);
+            }
+
+            if (this.MarkerMid != null)
+            {
+                SvgMarker marker = this.OwnerDocument.GetElementById<SvgMarker>(this.MarkerMid.ToString());
+                for (int i = 1; i <= path.PathPoints.Length - 2; i++)
+                    marker.RenderMarker(renderer, this, path.PathPoints[i], path.PathPoints[i - 1], path.PathPoints[i], path.PathPoints[i + 1]);
+            }
+
+            if (this.MarkerEnd != null)
+            {
+                SvgMarker marker = this.OwnerDocument.GetElementById<SvgMarker>(this.MarkerEnd.ToString());
+                marker.RenderMarker(renderer, this, path.PathPoints[path.PathPoints.Length - 1], path.PathPoints[path.PathPoints.Length - 2], path.PathPoints[path.PathPoints.Length - 1]);
+            }
+
+            return result;
         }
 
         public override RectangleF Bounds
