+++ Source/Css/SvgElementOps.cs
@@ -35,7 +35,7 @@ namespace Svg.Css
 
         public Selector<SvgElement> AttributeExists(NamespacePrefix prefix, string name)
         {
-            return nodes => nodes.Where(n => n.Attributes.ContainsKey(name) || n.CustomAttributes.ContainsKey(name));
+            return nodes => nodes.Where(n => n.ContainsAttribute(name));
         }
 
         public Selector<SvgElement> AttributeExact(NamespacePrefix prefix, string name, string value)
@@ -43,9 +43,7 @@ namespace Svg.Css
             return nodes => nodes.Where(n =>
             {
                 string val = null;
-                object oval = null;
-                return (n.CustomAttributes.TryGetValue(name, out val) && val == value) ||
-                    (n.Attributes.TryGetValue(name, out oval) && oval.ToString() == value);
+                return (n.TryGetAttribute(name, out val) && val == value);
             });
         }
 
@@ -54,9 +52,7 @@ namespace Svg.Css
             return nodes => nodes.Where(n =>
             {
                 string val = null;
-                object oval = null;
-                return (n.CustomAttributes.TryGetValue(name, out val) && val.Split(' ').Contains(value)) ||
-                    (n.Attributes.TryGetValue(name, out oval) && oval.ToString().Split(' ').Contains(value));
+                return (n.TryGetAttribute(name, out val) && val.Split(' ').Contains(value));
             });
         }
 
@@ -67,9 +63,7 @@ namespace Svg.Css
                  : (nodes => nodes.Where(n =>
                     {
                         string val = null;
-                        object oval = null;
-                        return (n.CustomAttributes.TryGetValue(name, out val) && val.Split('-').Contains(value)) ||
-                            (n.Attributes.TryGetValue(name, out oval) && oval.ToString().Split('-').Contains(value));
+                        return (n.TryGetAttribute(name, out val) && val.Split('-').Contains(value));
                     }));
         }
 
@@ -80,9 +74,7 @@ namespace Svg.Css
                  : (nodes => nodes.Where(n =>
                      {
                          string val = null;
-                         object oval = null;
-                         return (n.CustomAttributes.TryGetValue(name, out val) && val.StartsWith(value)) ||
-                             (n.Attributes.TryGetValue(name, out oval) && oval.ToString().StartsWith(value));
+                         return (n.TryGetAttribute(name, out val) && val.StartsWith(value));
                      }));
         }
 
@@ -93,9 +85,7 @@ namespace Svg.Css
                  : (nodes => nodes.Where(n =>
                  {
                      string val = null;
-                     object oval = null;
-                     return (n.CustomAttributes.TryGetValue(name, out val) && val.EndsWith(value)) ||
-                         (n.Attributes.TryGetValue(name, out oval) && oval.ToString().EndsWith(value));
+                     return (n.TryGetAttribute(name, out val) && val.EndsWith(value));
                  }));
         }
 
@@ -106,9 +96,7 @@ namespace Svg.Css
                  : (nodes => nodes.Where(n =>
                  {
                      string val = null;
-                     object oval = null;
-                     return (n.CustomAttributes.TryGetValue(name, out val) && val.Contains(value)) ||
-                         (n.Attributes.TryGetValue(name, out oval) && oval.ToString().Contains(value));
+                     return (n.TryGetAttribute(name, out val) && val.Contains(value));
                  }));
         }
